refactor(web): share basePath normalization across config and runtime
