+++ packages/web/next.config.js
@@ -1,20 +1,7 @@
 /** @type {import('next').NextConfig} */
+import { normalizeBasePath } from "./src/lib/base-path.js";
 
-export function normalizeBasePath(raw) {
-  if (raw == null) return "";
-
-  const trimmed = String(raw).trim();
-  if (!trimmed) return "";
-
-  const stripped = trimmed.replace(/^\/+|\/+$/g, "");
-  if (!stripped) {
-    throw new Error(
-      `Invalid NEXT_PUBLIC_BASE_PATH: "${raw}" must include a non-root path segment like "ao" or "/ao".`,
-    );
-  }
-
-  return `/${stripped}`;
-}
+export { normalizeBasePath };
 
 const normalizedBasePath = normalizeBasePath(process.env.NEXT_PUBLIC_BASE_PATH ?? "");
